+++ src/main/java/com/rideco/grocery/exceptions/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package com.rideco.grocery.exceptions;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Integer productId) {
+    public ProductNotFoundException(Long productId) {
         super(String.format("Product with Id %d not found", productId));
     }
 }
